Added integration testing for portfolio controller; fixed getId() in position domain to actually return an ID and not just null (this blew up the ITs); fixed some naming mismatch  issues such as averagePrice to avgPrice

+++ backend/src/main/java/com/portfolio/portfolio_manager/dto/PositionCreateRequest.java
@@ -8,7 +8,7 @@ import java.math.BigDecimal;
 
 public record PositionCreateRequest(
         @NotBlank(message = "symbol is required") String symbol,
-        @NotNull(message = "quantity is required") BigDecimal quantity,
+        @NotNull(message = "quantity is required") Integer quantity,
         @NotNull(message = "avgPrice is required")
         @DecimalMin(value = "0.0", inclusive = false, message = "avgPrice must be greater than 0")
         BigDecimal avgPrice
