+++ backend/src/main/java/com/portfolio/portfolio_manager/dto/PositionResponse.java
@@ -13,7 +13,7 @@ import java.util.UUID;
 public record PositionResponse(
     UUID id, 
     String symbol, 
-    BigDecimal quantity, 
+    Integer quantity, 
     BigDecimal avgPrice
 
  )  {}
